requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asttokens==3.0.0
beautifulsoup4==4.12.3
beets==2.2.0
beets-filetote==0.4.9
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
confuse==2.0.1
coverage==5.5
dacite==1.8.1
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
elastic-transport==8.17.0
elasticsearch==8.17.0
executing==2.2.0
eyed3==0.9.7
fastapi==0.103.2
filetype==1.2.0
getch==1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jaconv==0.4.0
jedi==0.19.2
jellyfish==1.1.3
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediafile==0.13.0
munkres==1.1.4
musicbrainzngs==0.7.1
mutagen==1.47.0
nest-asyncio==1.6.0
openai==1.59.9
packaging==24.2
paho-mqtt==2.1.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-json==0.1.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pykakasi==2.3.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytube==15.0.0
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.11.0
redis==5.2.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.6
spotdl==4.2.11
spotipy==2.25.1
stack-data==0.6.3
starlette==0.27.0
syncedlyrics==1.0.1
text-unidecode==1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.23.2
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
yt-dlp==2025.1.15
ytmusicapi==1.9.1
